requirements.txt
Flask
beautifulsoup4
requests
duckduckgo_search
python-dotenv